Moved `mix hex.docs` to `mix hex.publish docs`.
