+++ lib/mix/tasks/hex/publish.ex
@@ -88,18 +88,21 @@ defmodule Mix.Tasks.Hex.Publish do
     else
       case rest do
         ["package"] ->
-          create(build, auth, opts)
+          create_package(build, auth, opts)
+        ["docs"] ->
+          create_docs(auth, opts, args)
         _otherwise ->
           message = """
             Invalid arguments, expected one of:
               mix hex.publish package
+              mix hex.publish docs
             """
           Mix.raise message
       end
     end
   end
 
-  defp create(build, auth, opts) do
+  defp create_package(build, auth, opts) do
     meta = build[:meta]
     package = build[:package]
     exclude_deps = build[:exclude_deps]
@@ -115,6 +118,26 @@ defmodule Mix.Tasks.Hex.Publish do
     end
   end
 
+  defp create_docs(auth, opts, args) do
+    Mix.Project.get!
+
+    config  = Mix.Project.config
+    app     = config[:app]
+    version = config[:version]
+
+    Mix.Task.run("docs", args)
+
+    directory = docs_dir()
+
+    unless File.exists?("#{directory}/index.html") do
+      Mix.raise "File not found: #{directory}/index.html"
+    end
+
+    progress? = Keyword.get(opts, :progress, true)
+    tarball = build_tarball(app, version, directory)
+    send_tarball(app, version, tarball, auth, progress?)
+  end
+
   defp print_link_to_coc() do
     Hex.Shell.info "Before publishing, please read Hex Code of Conduct: https://hex.pm/docs/codeofconduct"
   end
@@ -126,8 +149,10 @@ defmodule Mix.Tasks.Hex.Publish do
     case Hex.API.Release.delete(meta[:name], version, auth) do
       {code, _} when code in 200..299 ->
         Hex.Shell.info("Reverted #{meta[:name]} v#{version}")
+        Hex.Shell.info("Reverted docs for #{meta[:name]} v#{version}")
       {code, body} ->
         Hex.Shell.error("Reverting #{meta[:name]} v#{version} failed")
+        Hex.Shell.error("Reverting docs for #{meta[:name]} v#{version} failed")
         Hex.Utils.print_error_result(code, body)
     end
   end
@@ -149,6 +174,62 @@ defmodule Mix.Tasks.Hex.Publish do
       {code, body} ->
         Hex.Shell.error("\nPushing #{meta[:name]} v#{meta[:version]} failed")
         Hex.Utils.print_error_result(code, body)
+    end
+  end
+
+  defp build_tarball(app, version, directory) do
+    tarball = "#{app}-#{version}-docs.tar.gz"
+    files = files(directory)
+    :ok = :erl_tar.create(tarball, files, [:compressed])
+    data = File.read!(tarball)
+
+    File.rm!(tarball)
+    data
+  end
+
+  defp send_tarball(app, version, tarball, auth, progress?) do
+    if progress? do
+      progress = Utils.progress(byte_size(tarball))
+    else
+      progress = Utils.progress(nil)
+    end
+
+    case Hex.API.ReleaseDocs.new(app, version, tarball, auth, progress) do
+      {code, _} when code in 200..299 ->
+        Hex.Shell.info ""
+        Hex.Shell.info "Published docs for #{app} v#{version}"
+        # TODO: Only print this URL if we use the default API URL
+        Hex.Shell.info "Hosted at #{Hex.Utils.hexdocs_url(app, version)}"
+      {code, _} when code == 404 ->
+        Hex.Shell.info ""
+        Hex.Shell.error "Pushing docs for #{app} v#{version} is not possible due to the package not be published"
+      {code, body} ->
+        Hex.Shell.info ""
+        Hex.Shell.error "Pushing docs for #{app} v#{version} failed"
+        Hex.Utils.print_error_result(code, body)
+    end
+  end
+
+  defp files(directory) do
+    "#{directory}/**"
+    |> Path.wildcard
+    |> Enum.filter(&File.regular?/1)
+    |> Enum.map(&{relative_path(&1, directory), File.read!(&1)})
+  end
+
+  defp relative_path(file, dir) do
+    Path.relative_to(file, dir)
+    |> String.to_char_list
+  end
+
+  defp docs_dir do
+    cond do
+      File.exists?("doc") ->
+        "doc"
+      File.exists?("docs") ->
+        "docs"
+      true ->
+        Mix.raise("Documentation could not be found. Please ensure documentation is in the doc/ or docs/ directory.")
     end
   end
 end
