+++ test/test_helper.exs
@@ -269,6 +269,7 @@ if :integration in ExUnit.configuration[:include] do
 
   Case.init_project("ex_doc", "0.0.1", [], pkg_meta, auth)
   Case.init_project("ex_doc", "0.0.1", [], pkg_meta, auth)
+  Case.init_project("ex_doc", "0.0.3", [], pkg_meta, auth)
   Case.init_project("ex_doc", "0.1.0", [], pkg_meta, auth)
   Case.init_project("postgrex", "0.2.1", [ex_doc: "~> 0.1.0"], %{}, auth)
   Case.init_project("postgrex", "0.2.0", [ex_doc: "0.0.1"], %{}, auth)
